refactor code- speaker_dialog prompt fix
