+++ python/notebooklm/.gitignore
@@ -10,3 +10,4 @@ structured_scene.json
 narrator_dialog.json
 speaker_dialog_voice.json
 sound_effects_timestamp.json
+generated/
