+++ python/notebooklm/docs/script_prompts_en.json
@@ -7,7 +7,7 @@
       },
       {
         "task_name": "Speaker Dialog Prompt for Parser",
-        "prompt": "The task is to analyze the given script and update the voice descriptions for each speaker based on the scene.\nThe initial voice descriptions provided a general overview of how each speaker sounds.\nThe updated script includes detailed voice descriptions for each line of dialogue,\nincorporating the emotional nuances and specific tones relevant to each scene.\nspeaker_1_description = \"\"\"\n        Emma's voice is expressive and dramatic in delivery, speaking at a moderately fast pace with a very close recording that almost has no background noise.\n    \"\"\"\nspeaker_2_description = \"\"\"\n        Leo's voice is deep and resonant, with a calm and authoritative tone. He speaks at a steady pace, ensuring clarity and precision in his delivery. The recording is clear with minimal background noise, providing a professional and engaging listening experience.\n    \"\"\""
+        "prompt": "The task is to analyze the given script and update the voice descriptions for each speaker based on the scene. Provide only scene title and dialogs for speakers in json .\nThe initial voice descriptions provided a general overview of how each speaker sounds.\nThe updated script includes detailed voice descriptions for each line of dialogue,\nincorporating the emotional nuances and specific tones relevant to each scene.\nspeaker_1_description = \"\"\"\n        Emma's voice is expressive and dramatic in delivery, speaking at a moderately fast pace with a very close recording that almost has no background noise.\n    \"\"\"\nspeaker_2_description = \"\"\"\n        Leo's voice is deep and resonant, with a calm and authoritative tone. He speaks at a steady pace, ensuring clarity and precision in his delivery. The recording is clear with minimal background noise, providing a professional and engaging listening experience.\n    \"\"\""
       },
       {
         "task_name": "Prompt parser for narrator description",
